backend-api/requirements.txt
Flask
Flask-CORS
tensorflow
numpy
scikit-learn
joblib
pillow
pybase64
opencv-python
gunicorn
pymongo[srv]
bcrypt
python-dotenv